requirements.txt
Django==2.2.5
gunicorn==19.9.0
django-heroku==0.3.1
djangorestframework==3.9.4
django-cors-headers==3.0.2
djangorestframework-jwt==1.11.0
model-mommy==1.6.0